refactor: split infra files
