+++ pdf_server/api/to_pdf.py
@@ -1,9 +1,9 @@
 from typing import Annotated, Optional
 
-from fastapi import APIRouter, File, Form, Response, UploadFile, status
+from fastapi import APIRouter, File, Response, UploadFile, status
 
 from consts import get_mimetype
-from infra import convert_bytes_to_pdf
+from infra.pdf import convert_bytes_to_pdf
 from utils import out_filename
 
 router = APIRouter(tags=["to-pdf"])
